Split pretty text field apart
