+++ lib/components/helpers/pretty-select-input.js
@@ -21,6 +21,20 @@ module.exports = React.createClass({
   },
 
   renderDefault: function () {
+    return this.props.config.createElement('pretty-text-helper', {
+      classes: this.props.classes,
+      tabIndex: this.props.field.tabIndex,
+      onChange: this.props.onChange,
+      onFocus: this.props.onFocus,
+      onBlur: this.props.onBlur,
+      value: this.props.getDisplayValue(),
+      selectedChoices: this.props.config.fieldSelectedReplaceChoices(this.props.field),
+      replaceChoices: this.props.config.fieldReplaceChoices(this.props.field),
+      onTagClick: this.onTagClick,
+      ref: 'textBox',
+      isEditable: this.props.isEnteringCustomValue
+    });
+
     if (this.props.isEnteringCustomValue) {
       return <input className={cx(_.extend({}, this.props.classes))} type="text" value={this.props.field.value}
                     onChange={this.props.onChange} onFocus={this.props.onFocus} onBlur={this.props.onBlur} />;
