+++ lib/default-config.js
@@ -113,6 +113,7 @@ module.exports = function (config) {
 
     createElement_ChoiceSectionHeader: React.createFactory(require('./components/helpers/choice-section-header')),
 
+    createElement_PrettyTextHelper: React.createFactory(require('./components/helpers/pretty-text-helper')),
 
     // Field default value factories. Give a default value for a specific type.
